% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarDemo.R
\name{plotVarDemo}
\alias{plotVarDemo}
\title{Plot Environmental Data}
\usage{
plotVarDemo(varURI, wsUrl = "www.opensilex.org/openSilexAPI/rest/")
}
\arguments{
\item{varURI}{uri of the variable to plot from the \code{\link{variableList}} function or the web service directly}

\item{wsUrl}{url of the webservice}
}
\description{
Demonstration function
}
\examples{
\donttest{
 connectToPHISWS(apiID="ws_private", 
     url = "www.opensilex.org/openSilexAPI/rest/", 
     "guest@opensilex.org",
     "guest")
 vars <- variableList()
 vars
 plotVar(vars$value[1])
}

}
